Fixed walls and color rotation and scaling
Walls and colors now are initiated with the transform attributes of the associated preview object. The scaling of the object is now dictated by the arrow keys and the rotation has been moved from the mouse to the "<" and ">" keys. The user can hold down these keys instead of pressing them to scale and rotate, reducing the number of key presses. Deletion of an object now includes use of the left mouse button.